requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.1
python-dotenv==1.0.0
mysqlclient==2.2.0
Werkzeug==3.0.1
email-validator==2.1.0
